Refactors tests, and removes RaSP from testing on dev
