+++ src/poli/tests/registry/chemistry/alphabet_qed.json
