+++ src/poli/tests/registry/proteins/3ned.pdb
