+++ tox.ini
@@ -35,7 +35,7 @@ description = check the code style with black
 deps =
     black
 commands =
-    black --check --diff {posargs:.}
+    black --check --diff .
 
 [testenv:poli-base-py39]
 description = run the tests with pytest on the base environment for poli
